Removed global private browsing support, which was removed from Firefox
Private browsing moe now works per window, not globally. This means
that it no longer makes sense to support it since all reporting and
logging the add-on does is on the browser level, not the window level.
Also, most private browsing support was incorrect anyway, since it
pertained remote data, not local data.